+++ imblearn/under_sampling/__init__.py
@@ -3,17 +3,20 @@ The :mod:`imblearn.under_sampling` provides methods to under-sample
 a dataset.
 """
 
-from .random_under_sampler import RandomUnderSampler
-from .tomek_links import TomekLinks
-from .cluster_centroids import ClusterCentroids
-from .nearmiss import NearMiss
-from .condensed_nearest_neighbour import CondensedNearestNeighbour
-from .one_sided_selection import OneSidedSelection
-from .neighbourhood_cleaning_rule import NeighbourhoodCleaningRule
-from .edited_nearest_neighbours import EditedNearestNeighbours
-from .edited_nearest_neighbours import RepeatedEditedNearestNeighbours
-from .edited_nearest_neighbours import AllKNN
-from .instance_hardness_threshold import InstanceHardnessThreshold
+from . import prototype_generation
+from .prototype_generation import ClusterCentroids
+
+from . import prototype_selection
+from .prototype_selection import RandomUnderSampler
+from .prototype_selection import TomekLinks
+from .prototype_selection import NearMiss
+from .prototype_selection import CondensedNearestNeighbour
+from .prototype_selection import OneSidedSelection
+from .prototype_selection import NeighbourhoodCleaningRule
+from .prototype_selection import EditedNearestNeighbours
+from .prototype_selection import RepeatedEditedNearestNeighbours
+from .prototype_selection import AllKNN
+from .prototype_selection import InstanceHardnessThreshold
 
 __all__ = [
     'RandomUnderSampler', 'TomekLinks', 'ClusterCentroids', 'NearMiss',
