+++ imblearn/under_sampling/prototype_selection/edited_nearest_neighbours.py
@@ -14,8 +14,8 @@ from collections import Counter
 import numpy as np
 from scipy.stats import mode
 
-from ..base import BaseMulticlassSampler
-from ..utils import check_neighbors_object
+from ...base import BaseMulticlassSampler
+from ...utils import check_neighbors_object
 
 SEL_KIND = ('all', 'mode')
