+++ imblearn/under_sampling/prototype_selection/neighbourhood_cleaning_rule.py
@@ -10,8 +10,8 @@ from collections import Counter
 
 import numpy as np
 
-from ..base import BaseMulticlassSampler
-from ..utils import check_neighbors_object
+from ...base import BaseMulticlassSampler
+from ...utils import check_neighbors_object
 
 
 class NeighbourhoodCleaningRule(BaseMulticlassSampler):
